refactor(runtime): move rendering/templates to runtime/ package (#158)
Phase 2 of #156. Establishes ``src/srunx/runtime/`` as the canonical
home for the submission-path layer.

Moves:
- ``src/srunx/rendering.py`` -> ``src/srunx/runtime/rendering.py``
- ``src/srunx/template.py`` -> ``src/srunx/runtime/templates.py``
- ``src/srunx/templates/base.slurm.jinja`` -> ``src/srunx/runtime/_jinja/base.slurm.jinja``

Old top-level modules stay as backward-compat shims re-exporting
every public + private name so external callers and monkeypatch-based
tests keep working. Internal consumers that hardcoded the old Jinja
path (client.py, mcp/server.py, 3 test files) were updated to point
at the new ``_jinja/`` location.

``tests/architecture/test_import_layers.py`` gains a ``runtime``
self-mapping and notes ``rendering`` / ``template`` as shim layers.
