+++ src/srunx/client.py
@@ -924,7 +924,7 @@ class Slurm:
 
     def _get_default_template(self) -> str:
         """Get the default job template path."""
-        return str(files("srunx.templates").joinpath("base.slurm.jinja"))
+        return str(files("srunx.runtime").joinpath("_jinja", "base.slurm.jinja"))
 
 
 # Convenience functions for backward compatibility
